Deprecate SnippetPreview atom
Inline in CssSnippetCard due to single-use.
