+++ src/components/atoms/_index.scss
@@ -3,4 +3,3 @@
 @import './collectionChip';
 @import './pageBackdrop';
 @import './pageTitle';
-@import './snippetPreview';
