+++ src/components/organisms/snippetCard/cssSnippetCard/index.jsx
@@ -1,9 +1,10 @@
+import { useEffect } from 'react';
 import PropTypes from 'typedefs/proptypes';
 import Card, { CardIcon } from 'components/atoms/card';
 import TagList from 'components/atoms/tagList';
 import CodeBlock from 'components/atoms/codeBlock';
 import Actions from 'components/molecules/actions';
-import SnippetPreview from 'components/atoms/snippetPreview';
+import literals from 'lang/en/client/common';
 
 const propTypes = {
   snippet: PropTypes.snippet,
@@ -14,6 +15,23 @@ const propTypes = {
  * Used for CSS snippets.
  */
 const SnippetCard = ({ snippet }) => {
+  useEffect(() => {
+    if (!snippet.code.js) return;
+    const s = document.createElement('script');
+    s.type = 'text/javascript';
+    s.async = true;
+    s.innerHTML = `function snippet_preview_js(){${snippet.code.js}};`;
+    document.body.appendChild(s);
+    try {
+      window['snippet_preview_js']();
+    } catch (e) {
+      /* istanbul ignore next */
+      console.warn(
+        'There is an issue with JavaScript execution on the snippet preview of this page.'
+      );
+    }
+  }, []);
+
   return (
     <Card className='snippet-card'>
       <div className='card-meta grid'>
@@ -28,12 +46,16 @@ const SnippetCard = ({ snippet }) => {
         dangerouslySetInnerHTML={{ __html: snippet.html.fullDescription }}
       />
       <div className='card-preview-content'>
-        <SnippetPreview
-          scopeId={snippet.id.slice(snippet.id.lastIndexOf('/') + 1)}
-          scopedCss={snippet.code.scopedCss}
-          htmlCode={snippet.code.html}
-          jsCode={snippet.code.js}
-        />
+        <h5 className='snippet-preview-title txt-150 fs-md'>
+          {literals.preview}
+        </h5>
+        <div
+          className='snippet-preview m-0 br-lg'
+          data-scope={snippet.id.slice(snippet.id.lastIndexOf('/') + 1)}
+        >
+          <style>{snippet.code.scopedCss}</style>
+          <div dangerouslySetInnerHTML={{ __html: snippet.code.html }} />
+        </div>
       </div>
       <div className='card-source-content'>
         <CodeBlock
